pkg/hostagent: split DNS pkg
